backend/requirements.txt
# Web Framework
django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Machine Learning & Cómputo
numpy==1.26.4
scikit-learn==1.4.1.post1
pandas==2.2.1

# Infraestructura & Cache
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.3
celery==5.3.6

# Utilidades
python-dotenv==1.0.1